refactor(modules): split adapters, examples, and adapter tests into separate Go modules
Move optional integrations into dedicated modules so the core package remains focused.
This keeps dependency resolution scoped to each use case and clarifies module boundaries.
