+++ adapter_tests/gorm_test.go
@@ -1,4 +1,4 @@
-package pgvector_test
+package adapter_tests_test
 
 import (
 	"math"
