+++ adapter_tests/stub.go
@@ -1,0 +1,5 @@
+package adapter_tests
+
+func stub() {
+	// Empty stub
+}
