+++ adapters/ent/functions.go
