+++ adapters/ent/test/generate.go
@@ -1,0 +1,3 @@
+package test
+
+//go:generate go run entgo.io/ent/cmd/ent generate --feature sql/execquery ./schema
