+++ adapters/pgx/go.mod
@@ -1,0 +1,17 @@
+module github.com/pgvector/pgvector-go/adapters/pgx
+
+go 1.23.0
+
+require (
+	github.com/jackc/pgx/v5 v5.7.2
+	github.com/pgvector/pgvector-go v0.3.0
+	github.com/x448/float16 v0.8.4
+)
+
+require (
+	github.com/jackc/pgpassfile v1.0.0 // indirect
+	github.com/jackc/pgservicefile v0.0.0-20240606120523-5a60cdf6a761 // indirect
+	github.com/stretchr/testify v1.8.2 // indirect
+	golang.org/x/crypto v0.36.0 // indirect
+	golang.org/x/text v0.23.0 // indirect
+)
