+++ examples/disco/main.go
@@ -7,7 +7,7 @@ import (
 	"github.com/ankane/disco-go"
 	"github.com/jackc/pgx/v5"
 	"github.com/pgvector/pgvector-go"
-	pgxvector "github.com/pgvector/pgvector-go/pgx"
+	pgxvector "github.com/pgvector/pgvector-go/adapters/pgx"
 )
 
 func main() {
