+++ examples/go.mod
@@ -1,0 +1,7 @@
+module github.com/pgvector/pgvector-go/examples
+
+go 1.23.0
+
+replace github.com/pgvector/pgvector-go/adapters/ent => ../adapters/ent
+
+replace github.com/pgvector/pgvector-go/adapters/pgx => ../adapters/pgx
