--- test/ent/generate.go
@@ -1,3 +1,0 @@
-package ent
-
-//go:generate go run -mod=mod entgo.io/ent/cmd/ent generate --feature sql/execquery ./schema
